+++ doers/src/file.rs
@@ -12,7 +12,7 @@ use std::fmt::Debug;
 use std::fs;
 use std::io::Write;
 use std::time::{SystemTime, UNIX_EPOCH};
-use util::file::{self, FileMetadata, UserOrGroup};
+use util::file::{self, FileMetadata, NameOrId};
 use util::http;
 
 // THINGS TO KNOW
@@ -39,10 +39,10 @@ pub struct DesiredFileState {
     pub from_struct: Option<Value>,
     pub from_url: Option<String>,
     pub from: Option<Utf8PathBuf>,
-    pub group: UserOrGroup,
+    pub group: NameOrId,
     pub ignore_pattern: Option<String>,
     pub mode: String,
-    pub owner: UserOrGroup,
+    pub owner: NameOrId,
     pub to_format: Option<String>,
     pub with_checksum: Option<String>,
     pub remote_content: RefCell<Option<Vec<u8>>>,
@@ -314,8 +314,8 @@ impl GurpFileEnsure {
                 fs::rename(&self.path, backup_target)?;
                 file::ensure_metadata(
                     FileMetadata {
-                        group: &UserOrGroup::Name("root".to_owned()),
-                        owner: &UserOrGroup::Name("root".to_owned()),
+                        group: &NameOrId::Name("root".to_owned()),
+                        owner: &NameOrId::Name("root".to_owned()),
                         mode: "0o0400",
                         path: backup_target,
                         changes: 1,
@@ -475,8 +475,8 @@ mod test {
                     from: None,
                     ignore_pattern: None,
                     mode: "0600".to_owned(),
-                    group: UserOrGroup::Name("root".to_owned()),
-                    owner: UserOrGroup::Name("dataperson".to_owned()),
+                    group: NameOrId::Name("root".to_owned()),
+                    owner: NameOrId::Name("dataperson".to_owned()),
                     to_format: None,
                     with_checksum: None,
                     remote_content: RefCell::new(None),
@@ -501,8 +501,8 @@ mod test {
                     with_checksum: Some("0123456789abcdef".to_owned()),
                     ignore_pattern: None,
                     mode: "0640".to_owned(),
-                    owner: UserOrGroup::Name("gibbus".to_owned()),
-                    group: UserOrGroup::Name("root".to_owned()),
+                    owner: NameOrId::Name("gibbus".to_owned()),
+                    group: NameOrId::Name("root".to_owned()),
                     to_format: None,
                     remote_content: RefCell::new(None),
                 }
@@ -693,8 +693,8 @@ mod test {
             path: temp_file.clone(),
             desired_state: DesiredFileState {
                 mode: "0755".to_owned(),
-                group: UserOrGroup::Name(my_group()),
-                owner: UserOrGroup::Name(my_user()),
+                group: NameOrId::Name(my_group()),
+                owner: NameOrId::Name(my_user()),
                 content: None,
                 ignore_pattern: None,
                 from: Some(fixture("file/binary-file")),
@@ -717,8 +717,8 @@ mod test {
             id: "IRRELEVANT".to_owned(),
             path: Utf8PathBuf::from("/does/not/matter"),
             desired_state: DesiredFileState {
-                group: UserOrGroup::Name(my_group()),
-                owner: UserOrGroup::Name(my_user()),
+                group: NameOrId::Name(my_group()),
+                owner: NameOrId::Name(my_user()),
                 mode: "2755".to_owned(),
                 content: None,
                 ignore_pattern: None,
@@ -738,8 +738,8 @@ mod test {
             id: "IRRELEVANT".to_owned(),
             path: Utf8PathBuf::from("/does/not/matter"),
             desired_state: DesiredFileState {
-                group: UserOrGroup::Name(my_group()),
-                owner: UserOrGroup::Name(my_user()),
+                group: NameOrId::Name(my_group()),
+                owner: NameOrId::Name(my_user()),
                 mode: "2755".to_owned(),
                 content: Some("content".to_owned()),
                 ignore_pattern: None,
@@ -759,8 +759,8 @@ mod test {
             id: "IRRELEVANT".to_owned(),
             path: Utf8PathBuf::from("/does/not/matter"),
             desired_state: DesiredFileState {
-                group: UserOrGroup::Name(my_group()),
-                owner: UserOrGroup::Name(my_user()),
+                group: NameOrId::Name(my_group()),
+                owner: NameOrId::Name(my_user()),
                 mode: "2755".to_owned(),
                 content: None,
                 ignore_pattern: None,
@@ -788,8 +788,8 @@ mod test {
             id: "IRRELEVANT".to_owned(),
             path: temp_file.clone(),
             desired_state: DesiredFileState {
-                group: UserOrGroup::Name(my_group()),
-                owner: UserOrGroup::Name(my_user()),
+                group: NameOrId::Name(my_group()),
+                owner: NameOrId::Name(my_user()),
                 mode: "2755".to_owned(),
                 content: None,
                 ignore_pattern: None,